+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.6.1
+# 0.7.0
 
 - add bool `callCubeDispose` in `CubeBuilder` `to decide call `dispose` or not in cube.
 - add method `listen` inner `Cube` to listen `ObservableValue`.